requirements.txt
streamlit>=1.35.0
pdfplumber>=0.11.0
langchain==0.2.16
langchain-core==0.2.38
langchain-groq==0.1.9
langchain-community==0.2.16
python-dotenv>=1.0.0
pydantic>=2.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0